COVID-19 data for total age/sex in Norway (2020 border).

This dataset contains COVID-19 data in Norway, from 2020 week 8 (2020-02-23) to 2022 week 45 (2022-11-13)

Usage

total_b2020

Format

granularity_time

Time granularity: day, isoyearweek

granularity_geo

Geographical granularity: nation, county, municip

location_code

Location code

border

Redistricting border: 2020

age

Age category: total

sex

Sex category: total

isoyear

Isoyear

isoweek

Isoweek

isoyearweek

Isoyearweek

season

Season

seasonweek

Season week

calyear

Calendar year

calmonth

Calendar month

date

Date

cases_by_testdate_n

Number of cases by date of PCR test

cases_by_testdate_vs_pop_pr100000

Number of cases by date of PCR test per 100.000 population

cases_by_regdate_n

Number of cases by date of registration in MSIS

cases_by_regdate_vs_pop_pr100000

Number of cases by date of registration in MSIS per 100.000 population

icu_admissions_n

Number of ICU admissions

hospital_admissions_main_cause_n

Number of hospital admissions with COVID-19 as the main cause

testevents_neg_n

Number of testing events (all tests of one person within a 7 day period are considered to be one testing event) that were negative

testevents_pos_n

Number of testing events (all tests of one person within a 7 day period are considered to be one testing event) that were positive

testevents_all_n

Number of testing events (all tests of one person within a 7 day period are considered to be one testing event)

testevents_pos_vs_all_pr100

Percent of testing events that were positive

vax_dose_1_by_vaxdate_n

Number of dose 1 vaccines delivered by date of vaccination

vax_dose_2_by_vaxdate_n

Number of dose 2 vaccines delivered by date of vaccination

vax_dose_3_by_vaxdate_n

Number of dose 3 vaccines delivered by date of vaccination

vax_dose_4_by_vaxdate_n

Number of dose 4 vaccines delivered by date of vaccination

vax_dose_1_by_vaxdate_sum0_999999_n

Cumulative number of dose 1 vaccines delivered by date of vaccination

vax_dose_2_by_vaxdate_sum0_999999_n

Cumulative number of dose 2 vaccines delivered by date of vaccination

vax_dose_3_by_vaxdate_sum0_999999_n

Cumulative number of dose 3 vaccines delivered by date of vaccination

vax_dose_4_by_vaxdate_sum0_999999_n

Cumulative number of dose 4 vaccines delivered by date of vaccination

vax_dose_1_by_regdate_n

Number of dose 1 vaccines delivered by date of registration in SYSVAK

vax_dose_2_by_regdate_n

Number of dose 2 vaccines delivered by date of registration in SYSVAK

vax_dose_3_by_regdate_n

Number of dose 3 vaccines delivered by date of registration in SYSVAK

vax_dose_4_by_regdate_n

Number of dose 4 vaccines delivered by date of registration in SYSVAK

pop_jan1_n

Population as per first of January

location_name

Location name (may not be unique)

location_name_description_nb

Location name description (is unique)

Examples

print(respiranor::total_b2020[1,])
#>    granularity_time granularity_geo country_iso3 location_code border   age
#> 1:             date          county          nor  county_nor03   2020 total
#>      sex isoyear isoweek isoyearweek    season seasonweek calyear calmonth
#> 1: total    2022      45     2022-45 2022/2023         16    2022       11
#>    calyearmonth       date cases_by_testdate_n
#> 1:     2022-M11 2022-11-13                  NA
#>    cases_by_testdate_vs_pop_pr100000 cases_by_regdate_n
#> 1:                                NA                 15
#>    cases_by_regdate_vs_pop_pr100000 icu_admissions_n
#> 1:                              2.1               NA
#>    hospital_admissions_main_cause_n testevents_neg_n testevents_pos_n
#> 1:                               NA               NA               NA
#>    testevents_all_n testevents_pos_vs_all_pr100 vax_dose_1_by_vaxdate_n
#> 1:               NA                          NA                      NA
#>    vax_dose_2_by_vaxdate_n vax_dose_3_by_vaxdate_n vax_dose_4_by_vaxdate_n
#> 1:                      NA                      NA                      NA
#>    vax_dose_1_by_vaxdate_sum0_999999_n vax_dose_2_by_vaxdate_sum0_999999_n
#> 1:                                  NA                                  NA
#>    vax_dose_3_by_vaxdate_sum0_999999_n vax_dose_4_by_vaxdate_sum0_999999_n
#> 1:                                  NA                                  NA
#>    vax_dose_1_by_regdate_n vax_dose_2_by_regdate_n vax_dose_3_by_regdate_n
#> 1:                       0                       1                      18
#>    vax_dose_4_by_regdate_n pop_jan1_n location_name
#> 1:                     132     699827          Oslo
#>    location_name_description_nb
#> 1:                 Oslo (fylke)